Towards autogenerated ast-nodes, visitors, and evaluation.
Notes...

The amount of boiler-plate and repetitive code patterns calls out for code generation.
It would be prefarable to have a code-base where only the interesting
code parts are manually written.

YAML currently seems more readable than JSON for the specification of the AST.
Parts of niels.json will probably be turned into variants.yaml
and the evaluation-code could probably be specified simpler in yaml as well.

Lets see where this goes...